backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.10.6
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
pdfplumber==0.11.5
pillow==10.4.0
pytesseract==0.3.10
